requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.3
zipp==3.18.2
